requirements.txt
fastapi
uvicorn
python-multipart
google-genai
faster-whisper
gTTS
pydantic
python-dotenv
